Initial commit for mobilenetv2 and configurator from other dev branch.

+++ benchmarks/mobilenetv2/hw/source/body/Residual.c
@@ -1,15 +1,20 @@
 #include "body_defines.h"
 
-void Residual(uint8_t enable, uint32_t length) {
+void Residual(uint64_t enable, uint64_t length) {
 	volatile dType_8u * streamIn0 = (dType_8u *)ResIn0;
 	volatile dType_8u * streamIn1 = (dType_8u *)ResIn1;
 	volatile dType_8u * streamOut = (dType_8u *)ResOut;
+	
+	// Checks that Sam hates™
+	// if(enable != RES_0_ENABLE) return;
+	// if(length != PW0_0_I_SIZE * PW0_0_I_SIZE * PW0_0_IC_SIZE) return;
+	
 	if (enable == 0) {
-		for(int i=0; i<length; i++) {
+		for(int i=0; i < length; i++) {
 			*streamOut = *streamIn0;
 		}
 	} else {
-		for(int i=0; i<length; i++) {
+		for(int i=0; i < length; i++) {
 			*streamOut = *streamIn0 + *streamIn1;
 		}
 	}
